Thoughts: So, the background threads in RobotC are interfering with existing threads. The reason I can safely say this is that there was no EndTimeSlice() in AccelIntegrate. This means that there MUST be another thread entering and reorganizing the stack priorities for threading. As a result, we can look again at the GyroIntegrate task. There are five lines in the functional task, plus one condition (at the time we last tested it). That means there's a 1 in 5 chance of being interrupted before the execution of any given line. Most importantly, I'm looking at the line: nPgmTime = lastIterTime. If the task is interrupted before this line is completed, the total time for the next and previous iterations will be VASTLY inaccurate.
As such, it makes sense to put hogCPU before both the gyroscope and accelerometer integrators. 

It occurred to me that the problem can be extended to the accelerometer. We were often overshooting our target, or being rather inaccurate in arriving at our target destination. If the accelerometer time function were skipping, this would explain the problem. However, because of the length of the accelerometer's intergrator, the chance of this happening is much less. This is VERY demonstrable in the error rate we are seeing. The temporal anomaly crossed with the 15-layer low pass filter means we are very likely causing errors by not using hogCPU.

The other problem this solves is the gyroscope's weird calibration numbers. 7.903 is a very strange number to receive as a calibration, and we have noted that it varies widely. Analyzing this trend, it occurs to me that the calibration value went up sharply as we added more functions, and as those tasks took longer, so too did the calibration errors in the gyroscope. 

This would mean the calibration in the gyroscope is almost entirely random, and falls most probably within certain bounds. Of course, hogCPU should fix this problem. That is the next test.